seems to work again

+++ AirlineDelays/src/readRecords.h
@@ -1,5 +1,5 @@
 int readRecord(char *line, int fieldNum);
-double readDelays(const char *filename, void *d, int fieldNum);
+unsigned long readDelays(const char *filename, void *d, int fieldNum);
 
 #define MAX_NUM_CHARS 2000
 #define MAX_NUM_VALUES 4000
@@ -13,7 +13,7 @@ typedef struct  {
     int min, max;
     int numValues;
     int observedMin, observedMax;
-    long values[MAX_NUM_VALUES + 1];
+    unsigned long values[MAX_NUM_VALUES + 1];
 } Table;
 
 Table* combineTables(Table **tables, int num, Table *out);
